libUI: Change function names a bit, allow blending of images that don't have an alpha channel
git-svn-id: svn://svn.freemyipod.org@527 2d50ea51-68f8-42ca-bb2e-f50f654680a9

+++ libs/ui/chooser.h
@@ -80,7 +80,7 @@ struct chooser_info
     const struct chooser_renderer* renderer;
     const void* rendererparams;
     const void* userparams;
-    int tickinterval;
+    long tickinterval;
     int itemcount;
     int defaultitem;
     const struct chooser_item items[];
